tests/wpt/mozilla/tests/mozilla/bluetooth/connect/get-same-gatt-server.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => Promise.all([device.gatt.connect(), device.gatt.connect()]))
    .then(gattServers => assert_true(gattServers[0], gattServers[1]))
    .catch(error => assert_unreached('Promise should have resolved. ' + error));
}, 'Multiple connects should return the same gatt object.');
</script>
